# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deokhugam API",
    "description": "Deokhugam API 문서",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/reviews": {
      "get": {
        "tags": [
          "리뷰 관리"
        ],
        "summary": "리뷰 목록 조회",
        "description": "검색 조건에 맞는 리뷰 목록을 조회합니다.",
        "operationId": "searchReviews",
        "parameters": [
          {
            "name": "reviewGetRequest",
            "in": "query",
            "description": "인기 리뷰 목록 조회용 DTO<br>\n- **period**: 랭킹 기간 (DAILY, WEEKLY, MONTHLY, ALL_TIME, 기본값: DAILY)<br>\n- **direction**: 정렬 방향 (ASC, DESC, 기본값: ASC)<br>\n- **cursor**: 커서 페이지네이션을 위한 커서<br>\n- **after**: createdAt 기준 보조 커서<br>\n- **limit**: 페이지 크기 (기본값: 50)\n",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReviewGetRequest"
            }
          },
          {
            "name": "Deokhugam-Request-User-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "리뷰 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (정렬 방향 오류, 페이지네이션 파라미터 오류, 리뷰 ID 누락)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "리뷰 관리"
        ],
        "summary": "리뷰 등록",
        "description": "새로운 리뷰를 등록합니다.",
        "operationId": "createReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "리뷰 등록 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 작성된 리뷰 존재",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (입력값 검증 실패)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리뷰 정보 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      }
    },
...
    "/api/reviews/{reviewId}": {
      "get": {
        "tags": [
          "리뷰 관리"
        ],
        "summary": "리뷰 상세 정보 조회",
        "description": "리뷰 ID로 상세 정보를 조회합니다.",
        "operationId": "getReview",
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "리뷰 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Deokhugam-Request-User-ID",
            "in": "header",
            "description": "요청자 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "리뷰 정보 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (요청자 ID 누락)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리뷰 정보 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "리뷰 관리"
        ],
        "summary": "리뷰 논리 삭제",
        "description": "본인이 작성한 리뷰을 논리적으로 삭제합니다.",
        "operationId": "deleteReview",
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "리뷰 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Deokhugam-Request-User-ID",
            "in": "header",
            "description": "요청자 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "리뷰 삭제 성공"
          },
          "400": {
            "description": "잘못된 요청 (요청자 ID 누락)"
          },
          "404": {
            "description": "리뷰 정보 없음"
          },
          "403": {
            "description": "리뷰 삭제 권한 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "리뷰 관리"
        ],
        "summary": "리뷰 수정",
        "description": "본인이 작성한 리뷰을 수정합니다.",
        "operationId": "updateReview",
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "리뷰 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Deokhugam-Request-User-ID",
            "in": "header",
            "description": "요청자 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "리뷰 수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (입력값 검증 실패, 요청자 ID 누락)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리뷰 정보 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "리뷰 수정 권한 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      }
    },
...
    "/api/reviews/popular": {
      "get": {
        "tags": [
          "리뷰 관리"
        ],
        "summary": "인기 리뷰 목록 조회",
        "description": "기간별 인기 리뷰 목록을 조회합니다",
        "operationId": "getPopularReviews",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "랭킹 기간",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAILY",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY",
                "ALL_TIME"
              ]
            }
          },
          {
            "name": "direction",
            "in": "query",
            "description": "정렬 방향",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "커서 페이지네이션 커서",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "보조 커서(createdAt)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지 크기",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "인기 리뷰 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PopularReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (랭킹 기간 오류, 정렬 방향 오류)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/books/{bookId}/hard": {
      "delete": {
        "tags": [
          "도서 관리"
        ],
        "summary": "도서 물리 삭제",
        "description": "도서를 물리적으로 삭제합니다.",
        "operationId": "hardDelete",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "도서 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "9a4d6e2b-8f53-4c8c-a7c7-21d5079b8b37"
          }
        ],
        "responses": {
          "204": {
            "description": "도서 삭제 성공"
          },
          "404": {
            "description": "도서 정보 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ReviewCreateRequest": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "minLength": 1
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          }
        },
        "required": [
          "bookId",
          "content",
          "userId"
        ]
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "bookTitle": {
            "type": "string"
          },
          "bookThumbnailUrl": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userNickname": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "commentCount": {
            "type": "integer",
            "format": "int64"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ContentDisposition": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "charset": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "deprecated": true
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time",
            "deprecated": true
          },
          "readDate": {
            "type": "string",
            "format": "date-time",
            "deprecated": true
          },
          "formData": {
            "type": "boolean"
          },
          "attachment": {
            "type": "boolean"
          },
          "inline": {
            "type": "boolean"
          }
        }
      },
      "DefaultHttpStatusCode": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "titleMessageCode": {
            "type": "string"
          },
          "detailMessageCode": {
            "type": "string"
          },
          "detailMessageArguments": {
            "type": "array",
            "items": {}
          },
          "typeMessageCode": {
            "type": "string"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "properties": {
              "allow": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HttpMethod"
                },
                "uniqueItems": true
              },
              "cacheControl": {
                "type": "string"
              },
              "contentLanguage": {
                "type": "string"
              },
              "etag": {
                "type": "string"
              },
              "acceptLanguage": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "range": {
                      "type": "string"
                    },
                    "weight": {
                      "type": "number",
                      "format": "double"
                    }
                  }
                }
              },
              "basicAuth": {
                "type": "string",
                "writeOnly": true
              },
              "accept": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MediaType"
                }
              },
              "acceptPatch": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MediaType"
                }
              },
              "bearerAuth": {
                "type": "string",
                "writeOnly": true
              },
              "connection": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "acceptCharset": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "expires": {
                "type": "integer",
                "format": "int64"
              },
              "pragma": {
                "type": "string"
              },
              "origin": {
                "type": "string"
              },
              "ifNoneMatch": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "ifMatch": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "upgrade": {
                "type": "string"
              },
              "vary": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "range": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HttpRange"
                }
              },
              "contentType": {
                "$ref": "#/components/schemas/MediaType"
              },
              "ifModifiedSince": {
                "type": "integer",
                "format": "int64"
              },
              "contentLength": {
                "type": "integer",
                "format": "int64"
              },
              "empty": {
                "type": "boolean"
              },
              "location": {
                "type": "string",
                "format": "uri"
              },
              "host": {
                "type": "object",
                "properties": {
                  "hostString": {
                    "type": "string"
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "hostAddress": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string",
                        "format": "byte"
                      },
                      "hostName": {
                        "type": "string"
                      },
                      "linkLocalAddress": {
                        "type": "boolean"
                      },
                      "multicastAddress": {
                        "type": "boolean"
                      },
                      "anyLocalAddress": {
                        "type": "boolean"
                      },
                      "loopbackAddress": {
                        "type": "boolean"
                      },
                      "siteLocalAddress": {
                        "type": "boolean"
                      },
                      "mcglobal": {
                        "type": "boolean"
                      },
                      "mcnodeLocal": {
                        "type": "boolean"
                      },
                      "mclinkLocal": {
                        "type": "boolean"
                      },
                      "mcsiteLocal": {
                        "type": "boolean"
                      },
                      "mcorgLocal": {
                        "type": "boolean"
                      },
                      "canonicalHostName": {
                        "type": "string"
                      }
                    }
                  },
                  "port": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "unresolved": {
                    "type": "boolean"
                  },
                  "hostName": {
                    "type": "string"
                  }
                }
              },
              "all": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "writeOnly": true
              },
              "lastModified": {
                "type": "integer",
                "format": "int64"
              },
              "date": {
                "type": "integer",
                "format": "int64"
              },
              "accessControlRequestMethod": {
                "$ref": "#/components/schemas/HttpMethod"
              },
              "accessControlExposeHeaders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "accessControlRequestHeaders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "accessControlAllowCredentials": {
                "type": "boolean"
              },
              "accessControlAllowMethods": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HttpMethod"
                }
              },
              "accessControlAllowOrigin": {
                "type": "string"
              },
              "ifUnmodifiedSince": {
                "type": "integer",
                "format": "int64"
              },
              "accessControlMaxAge": {
                "type": "integer",
                "format": "int64"
              },
              "contentDisposition": {
                "$ref": "#/components/schemas/ContentDisposition"
              },
              "acceptLanguageAsLocales": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "accessControlAllowHeaders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "statusCode": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DefaultHttpStatusCode"
              },
              {
                "$ref": "#/components/schemas/HttpStatus"
              }
            ]
          },
          "body": {
            "$ref": "#/components/schemas/ProblemDetail"
          }
        }
      },
      "HttpMethod": {},
      "HttpRange": {},
      "HttpStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        ],
        "enum": [
          "100 CONTINUE",
          "101 SWITCHING_PROTOCOLS",
          "102 PROCESSING",
          "103 EARLY_HINTS",
          "103 CHECKPOINT",
          "200 OK",
          "201 CREATED",
          "202 ACCEPTED",
          "203 NON_AUTHORITATIVE_INFORMATION",
          "204 NO_CONTENT",
          "205 RESET_CONTENT",
          "206 PARTIAL_CONTENT",
          "207 MULTI_STATUS",
          "208 ALREADY_REPORTED",
          "226 IM_USED",
          "300 MULTIPLE_CHOICES",
          "301 MOVED_PERMANENTLY",
          "302 FOUND",
          "302 MOVED_TEMPORARILY",
          "303 SEE_OTHER",
          "304 NOT_MODIFIED",
          "305 USE_PROXY",
          "307 TEMPORARY_REDIRECT",
          "308 PERMANENT_REDIRECT",
          "400 BAD_REQUEST",
          "401 UNAUTHORIZED",
          "402 PAYMENT_REQUIRED",
          "403 FORBIDDEN",
          "404 NOT_FOUND",
          "405 METHOD_NOT_ALLOWED",
          "406 NOT_ACCEPTABLE",
          "407 PROXY_AUTHENTICATION_REQUIRED",
          "408 REQUEST_TIMEOUT",
          "409 CONFLICT",
          "410 GONE",
          "411 LENGTH_REQUIRED",
          "412 PRECONDITION_FAILED",
          "413 PAYLOAD_TOO_LARGE",
          "413 REQUEST_ENTITY_TOO_LARGE",
          "414 URI_TOO_LONG",
          "414 REQUEST_URI_TOO_LONG",
          "415 UNSUPPORTED_MEDIA_TYPE",
          "416 REQUESTED_RANGE_NOT_SATISFIABLE",
          "417 EXPECTATION_FAILED",
          "418 I_AM_A_TEAPOT",
          "419 INSUFFICIENT_SPACE_ON_RESOURCE",
          "420 METHOD_FAILURE",
          "421 DESTINATION_LOCKED",
          "422 UNPROCESSABLE_ENTITY",
          "423 LOCKED",
          "424 FAILED_DEPENDENCY",
          "425 TOO_EARLY",
          "426 UPGRADE_REQUIRED",
          "428 PRECONDITION_REQUIRED",
          "429 TOO_MANY_REQUESTS",
          "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
          "451 UNAVAILABLE_FOR_LEGAL_REASONS",
          "500 INTERNAL_SERVER_ERROR",
          "501 NOT_IMPLEMENTED",
          "502 BAD_GATEWAY",
          "503 SERVICE_UNAVAILABLE",
          "504 GATEWAY_TIMEOUT",
          "505 HTTP_VERSION_NOT_SUPPORTED",
          "506 VARIANT_ALSO_NEGOTIATES",
          "507 INSUFFICIENT_STORAGE",
          "508 LOOP_DETECTED",
          "509 BANDWIDTH_LIMIT_EXCEEDED",
          "510 NOT_EXTENDED",
          "511 NETWORK_AUTHENTICATION_REQUIRED"
        ]
      },
...
      "MediaType": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "subtype": {
            "type": "string"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "qualityValue": {
            "type": "number",
            "format": "double"
          },
          "charset": {
            "type": "string"
          },
          "concrete": {
            "type": "boolean"
          },
          "wildcardType": {
            "type": "boolean"
          },
          "wildcardSubtype": {
            "type": "boolean"
          },
          "subtypeSuffix": {
            "type": "string"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "ReviewUpdateRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          }
        },
        "required": [
          "content"
        ]
      },
...
      "PopularReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reviewId": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "bookTitle": {
            "type": "string"
          },
          "bookThumbnailUrl": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userNickname": {
            "type": "string"
          },
          "reviewContent": {
            "type": "string"
          },
          "reviewRating": {
            "type": "number",
            "format": "double"
          },
          "period": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "MONTHLY",
              "ALL_TIME"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "rank": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "commentCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ReviewGetRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "keyword": {
            "type": "string"
          },
          "cursor": {},
          "after": {
            "type": "string",
            "format": "date-time"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "CustomHeaderAuth": {
        "type": "apiKey",
        "name": "Deokhugam-Request-User-ID",
        "in": "header"
      }
    }
  }
}